chore(src): reorganize Modal elements
