+++ src/helix-ui/elements/hx-modal/_shadow.less
@@ -1,4 +1,4 @@
-@import "./HXElement";
+@import "../HXElement";
 @import (reference) "components/buttons";
 
 #hxBackdrop {
